requirements.txt
pillow>=9.0.0
pystray>=0.19.0 